+++ mjx/mujoco/mjx/_src/collision_driver_test.py
@@ -253,7 +253,6 @@ class CapsuleCollisionTest(parameterized.TestCase):
     self.assertGreater(c.dist[1], 0)
     # extract the contact point with penetration
     c = jax.tree_map(lambda x: jp.take(x, 0, axis=0)[None], dx.contact)
-    c = c.replace(dim=c.dim[np.array([0])])
     for field in dataclasses.fields(Contact):
       _assert_attr_eq(c, d.contact, field.name, 'capsule_convex_edge', 1e-4)
 
@@ -281,7 +280,6 @@ class ConvexTest(absltest.TestCase):
     np.testing.assert_array_less(-dx.contact.dist[2:], 0)
     # extract the contact points with penetration
     c = jax.tree_map(lambda x: jp.take(x, jp.array([0, 1]), axis=0), dx.contact)
-    c = c.replace(dim=c.dim[np.array([0, 1])])
     for field in dataclasses.fields(Contact):
       _assert_attr_eq(c, d.contact, field.name, 'box_plane', 1e-2)
 
@@ -339,7 +337,6 @@ class ConvexTest(absltest.TestCase):
     np.testing.assert_array_less(-dx.contact.dist[1:], 0)
     # extract the contact point with penetration
     c = jax.tree_map(lambda x: jp.take(x, 0, axis=0)[None], dx.contact)
-    c = c.replace(dim=c.dim[np.array([0])])
     for field in dataclasses.fields(Contact):
       _assert_attr_eq(c, d.contact, field.name, 'box_box_edge', 1e-2)
 
@@ -517,8 +514,8 @@ class TopKContactTest(absltest.TestCase):
     dx_all = collision_jit_fn(mx_all, dx)
     dx_top_k = collision_jit_fn(mx_top_k, dx)
 
-    self.assertEqual(dx_all.ncon, 3)
-    self.assertEqual(dx_top_k.ncon, 2)
+    self.assertEqual(dx_all.contact.dist.shape, (3,))
+    self.assertEqual(dx_top_k.contact.dist.shape, (2,))
 
 
 if __name__ == '__main__':