+++ mjx/mujoco/mjx/_src/device.py
@@ -25,6 +25,7 @@ import mujoco
 from mujoco.mjx._src import collision_driver
 from mujoco.mjx._src import mesh
 from mujoco.mjx._src import types
+import numpy as np
 
 _MJ_TYPE_ATTR = {
     mujoco.mjtBias: (mujoco.MjModel.actuator_biastype,),
@@ -68,7 +69,6 @@ _TRANSFORMS = {
     (types.Data, 'xmat'): lambda x: x.reshape(x.shape[:-1] + (3, 3)),
     (types.Data, 'geom_xmat'): lambda x: x.reshape(x.shape[:-1] + (3, 3)),
     (types.Data, 'site_xmat'): lambda x: x.reshape(x.shape[:-1] + (3, 3)),
-    (types.Model, 'actuator_trnid'): lambda x: x[:, 0],
     (types.Contact, 'frame'): (
         lambda x: x.reshape(x.shape[:-1] + (3, 3))  # pylint: disable=g-long-lambda
         if x is not None and x.shape[0] else jp.zeros((0, 3, 3))
@@ -265,9 +265,16 @@ def device_get_into(result, value):
 
   else:
     if isinstance(result, mujoco.MjData):
+      ncon = value.contact.dist.shape[0]
+      nefc = value.efc_J.shape[0]
       mujoco._functions._realloc_con_efc(  # pylint: disable=protected-access
-          result, ncon=value.ncon, nefc=value.nefc
+          result, ncon=ncon, nefc=nefc
       )
+      result.ncon = ncon
+      result.nefc = nefc
+      efc_start = nefc - ncon * 4
+      result.contact.efc_address[:] = np.arange(efc_start, nefc, 4)
+      result.contact.dim[:] = 3
 
     for f in dataclasses.fields(value):  # type: ignore
       if (type(value), f.name) in _DERIVED: