+++ mjx/mujoco/mjx/_src/scan_test.py
@@ -210,15 +210,16 @@ class ScanTest(absltest.TestCase):
         m, fn, 'ujqva', 'ujqva', *args, group_by='u'
     )
 
+    actuator_trnid = m.actuator_trnid[:, 0]
     np.testing.assert_array_equal(gear, m.actuator_gear)
-    np.testing.assert_array_equal(jnt_typ, m.jnt_type[m.actuator_trnid])
+    np.testing.assert_array_equal(jnt_typ, m.jnt_type[actuator_trnid])
     np.testing.assert_array_equal(act, jp.array([1.4, 1.1]))
     expected_vadr = np.concatenate(
-        [np.nonzero(m.dof_jntid == trnid)[0] for trnid in m.actuator_trnid]
+        [np.nonzero(m.dof_jntid == trnid)[0] for trnid in actuator_trnid]
     )
     np.testing.assert_array_equal(vadr, expected_vadr)
     expected_qadr = np.concatenate(
-        [np.nonzero(scan._q_jointid(m) == i)[0] for i in m.actuator_trnid]
+        [np.nonzero(scan._q_jointid(m) == i)[0] for i in actuator_trnid]
     )
     np.testing.assert_array_equal(qadr, expected_qadr)